+++ src/main/java/de/erichseifert/gral/plots/areas/DefaultAreaRenderer2D.java
@@ -59,7 +59,7 @@ public class DefaultAreaRenderer2D extends AbstractAreaRenderer {
 		double axisYMin = axis.getMin().doubleValue();
 		double axisYMax = axis.getMax().doubleValue();
 		double axisYOrigin = MathUtils.limit(0.0, axisYMin, axisYMax);
-		double posYOrigin = axisRenderer.getPosition(axis, axisYOrigin, true, false).get(1).doubleValue();
+		double posYOrigin = axisRenderer.getPosition(axis, axisYOrigin, true, false).get(1);
 
 		GeneralPath path = new GeneralPath();
